refactor: split user managers and remove phone field in Profile
